创建离线跑批接口修改

+++ src/main/java/cn/realai/online/core/service/BatchRecordService.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 public interface BatchRecordService {
 
-	Long insert(BatchRecord batchRecord);
+	Integer insert(BatchRecord batchRecord);
 
 	Integer delete(List<Long> idList);
